# WebAtrio/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Persons, Companies, and Jobs API",
    "version": "1.0.0",
    "description": "API for managing persons, companies, and jobs"
  },
  "paths": {
    "/persons": {
      "get": {
        "summary": "Get all persons",
        "tags": [ "Person" ],
        "responses": {
          "200": {
            "description": "A list of persons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Person" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new person",
        "tags": [ "Person" ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Person created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Person" }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/persons/{person_id}": {
      "get": {
        "summary": "Get a person by ID",
        "tags": [ "Person" ],
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Person details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Person" }
              }
            }
          },
          "404": {
            "description": "Person not found"
          }
        }
      },
      "put": {
        "summary": "Update a person by ID",
        "tags": [ "Person" ],
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Person updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Person" }
              }
            }
          },
          "404": {
            "description": "Person not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a person by ID",
        "tags": [ "Person" ],
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Person deleted successfully"
          },
          "404": {
            "description": "Person not found"
          }
        }
      }
    },
    "/persons/{person_id}/jobs": {
      "get": {
        "summary": "Get a person by ID",
        "tags": [ "Person" ],
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The start date for filtering jobs (optional).",
            "in": "query",
            "name": "startDate",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2024-01-01"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2024-12-31"
            },
            "description": "The end date for filtering jobs (optional)."
          }
        ],
        "responses": {
          "200": {
            "description": "Person details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Person" }
              }
            }
          },
          "404": {
            "description": "Person not found"
          }
        }
      }
    },
    "/companies": {
      "get": {
        "summary": "Get all companies",
        "tags": [ "Company" ],
        "responses": {
          "200": {
            "description": "A list of companies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Company" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new company",
        "tags": [ "Company" ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Company"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Company created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Company" }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/companies/{company_id}": {
      "get": {
        "summary": "Get a company by ID",
        "tags": [ "Company" ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Company" }
              }
            }
          },
          "404": {
            "description": "Company not found"
          }
        }
      },
      "put": {
        "summary": "Update a company by ID",
        "tags": [ "Company" ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Company"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Company" }
              }
            }
          },
          "404": {
            "description": "Company not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a company by ID",
        "tags": [ "Company" ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Company deleted successfully"
          },
          "404": {
            "description": "Company not found"
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "summary": "Get all jobs",
        "tags": [ "Job" ],
        "responses": {
          "200": {
            "description": "A list of jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Job" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new job",
        "tags": [ "Job" ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Job"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Job" }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "Get a job by ID",
        "tags": [ "Job" ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Job" }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "put": {
        "summary": "Update a job by ID",
        "tags": [ "Job" ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Job"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Job" }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a job by ID",
        "tags": [ "Job" ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Job deleted successfully"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Person": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          }
        },
        "required": [ "firstName", "lastName", "birthDate" ]
      },
      "Company": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [ "name" ]
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "position": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable":  true
          },
          "companyId": {
            "type": "integer"
          },
          "personId": {
            "type": "integer"
          }
        },
        "required": [ "position", "startDate", "companyId", "personId" ]
      }
    }
  }
}